// src/types.ts
export type ToolCategory = 
  | '3D'
  | 'Automation'
  | 'Business'
  | 'Collaboration'
  | 'Communication'
  | 'Creativity'
  | 'Data Science'
  | 'Design'
  | 'Education'
  | 'Electronics'
  | 'Freelancing'
  | 'Gamification'
  | 'Games'
  | 'Language Learning'
  | 'Library'
  | 'Lifestyle'
  | 'Math'
  | 'Memory'
  | 'Online Learning'
  | 'Other'
  | 'Productivity'
  | 'Programming'
  | 'Project Management'
  | 'Reading'
  | 'Research'
  | 'Science'
  | 'Studying'
  | 'Teaching'
  | 'Technology'
  | 'Test Prep'
  | 'Video'
  | 'Writing';

export type ToolPricing = 
  | 'Free'
  | 'Freemium'
  | 'Paid'
  | 'Subscription';